pyproject.toml
[project]
name = "dino-rag-clip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.3.5",
    "llama-index>=0.14.10",
    "llama-index-core>=0.14.10",
    "llama-index-embeddings-gemini>=0.4.1",
    "llama-index-embeddings-google-genai>=0.3.1",
    "llama-index-llms-gemini>=0.6.1",
    "llama-index-llms-google-genai>=0.8.0",
    "llama-index-readers-file>=0.5.5",
    "llama-index-vector-stores-chroma>=0.5.3",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "pydantic-ai-slim[google]>=1.28.0",
    "pysqlite3-binary>=0.5.4.post2",
    "python-dotenv>=1.2.1",
]
